finishing touch css updates etc

+++ checkout/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect, reverse
+from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.conf import settings
 from .forms import OrderForm
@@ -10,7 +10,7 @@ from cart.contexts import cart_contents
 from django.views.decorators.http import require_POST
 from django.http import HttpResponse
 import stripe
-
+import json
 
 @require_POST
 def cache_checkout_data(request):
@@ -18,20 +18,33 @@ def cache_checkout_data(request):
         pid = request.POST.get('client_secret').split('_secret')[0]
         stripe.api_key = settings.STRIPE_SECRET_KEY
         stripe.PaymentIntent.modify(pid, metadata={
-            'bag': json.dumps(request.session.get('bag', {})),
+            'cart': json.dumps(request.session.get('cart', {})),
             'save_info': request.POST.get('save_info'),
-            'username': request.user,
+            'username': request.user.username,
         })
         return HttpResponse(status=200)
     except Exception as e:
-        messages.error(request, ('Sorry, your payment cannot be '
-                                 'processed right now. Please try '
-                                 'again later.'))
+        messages.error(request, 'Sorry, your payment cannot be processed right now. Please try again later.')
         return HttpResponse(content=e, status=400)
 
 def checkout(request):
     stripe_public_key = settings.STRIPE_PUBLIC_KEY
     stripe_secret_key = settings.STRIPE_SECRET_KEY
+    order_number = None
+
+    try:
+        cart = request.session.get('cart', {})
+        current_cart = cart_contents(request)
+        total = current_cart['grand_total']
+        stripe_total = round(total * 100)
+        stripe.api_key = stripe_secret_key
+        intent = stripe.PaymentIntent.create(
+            amount=stripe_total,
+            currency=settings.STRIPE_CURRENCY,
+        )
+    except Exception as e:
+        messages.error(request, f"An error occurred while creating PaymentIntent: {str(e)}")
+        intent = None
 
     if request.method == 'POST':
         cart = request.session.get('cart', {})
@@ -49,64 +62,39 @@ def checkout(request):
 
         order_form = OrderForm(form_data)
         if order_form.is_valid():
-            order = order_form.save(commit=False)
             pid = request.POST.get('client_secret').split('_secret')[0]
+            order = order_form.save(commit=False)
             order.stripe_pid = pid
             order.original_bag = json.dumps(cart)
             order.save()
-            for item_id, item_data in cart.items():
-                try:
-                    product = Product.objects.get(id=item_id)
-                    if isinstance(item_data, int):
+
+            order_number = order.order_number
+
+            if isinstance(cart, dict):
+                for item_id, item_data in cart.items():
+                    try:
+                        product = Product.objects.get(id=item_id)
+                        quantity = item_data if isinstance(item_data, int) else item_data['quantity']
                         order_line_item = OrderLineItem(
                             order=order,
                             product=product,
-                            quantity=item_data,
+                            quantity=quantity,
                         )
                         order_line_item.save()
-                    else:
-                        for size, quantity in item_data['items_by_size'].items():
-                            order_line_item = OrderLineItem(
-                                order=order,
-                                product=product,
-                                quantity=quantity,
-                                product_size=size,
-                            )
-                            order_line_item.save()
-                except Product.DoesNotExist:
-                    messages.error(request, (
-                        "One of the products in your bag wasn't "
-                        "found in our database. "
-                        "Please call us for assistance!")
-                    )
-                    order.delete()
-                    return redirect(reverse('view_cart'))
+                    except Product.DoesNotExist:
+                        messages.error(request, "One of the products in your cart wasn't found in our database. Please call us for assistance!")
+                        order.delete()
+                        return redirect(reverse('view_cart'))
 
-            # Save the info to the user's profile if all is well
             request.session['save_info'] = 'save-info' in request.POST
-            return redirect(reverse('checkout_success',
-                                    args=[order.order_number]))
+            return redirect(reverse('checkout_success', args=[order_number]))
         else:
-            messages.error(request, ('There was an error with your form. '
-                                     'Please double check your information.'))
+            messages.error(request, 'There was an error with your form. Please double check your information.')
     else:
-        cart = request.session.get('cart', {})
         if not cart:
-            messages.error(request,
-                           "There's nothing in your cart at the moment")
+            messages.error(request, "There's nothing in your cart at the moment")
             return redirect(reverse('shop'))
 
-        current_cart = cart_contents(request)
-        total = current_cart['grand_total']
-        stripe_total = round(total * 100)
-        stripe.api_key = stripe_secret_key
-        intent = stripe.PaymentIntent.create(
-            amount=stripe_total,
-            currency=settings.STRIPE_CURRENCY,
-        )
-
-        # Attempt to prefill the form with any info
-        # the user maintains in their profile
         if request.user.is_authenticated:
             try:
                 profile = UserProfile.objects.get(user=request.user)
@@ -127,29 +115,50 @@ def checkout(request):
             order_form = OrderForm()
 
     if not stripe_public_key:
-        messages.warning(request, ('Stripe public key is missing. '
-                                   'Did you forget to set it in '
-                                   'your environment?'))
+        messages.warning(request, 'Stripe public key is missing. Did you forget to set it in your environment?')
 
-    template = 'checkout/checkout.html'
     context = {
         'order_form': order_form,
         'stripe_public_key': stripe_public_key,
-        'client_secret': intent.client_secret,
     }
+    if intent:
+        context['client_secret'] = intent.client_secret
 
-    return render(request, template, context)
+    return render(request, 'checkout/checkout.html', context)
+
 
 def checkout_success(request, order_number):
-    try:
-        order = Order.objects.get(order_number=order_number)
-        messages.success(request, f'Order successfully processed! Your order number is {order_number}. '
-                                  f'A confirmation email will be sent to {order.email}.')
+    if order_number:
+        order = get_object_or_404(Order, order_number=order_number)
+        print(f"Order retrieved successfully with order number: {order_number}")
+
+        if request.user.is_authenticated:
+            profile = UserProfile.objects.get(user=request.user)
+            order.user_profile = profile
+            order.save()
+
+            if request.session['save_info']:
+                profile_data = {
+                    'default_phone_number': order.phone_number,
+                    'default_country': order.country,
+                    'default_postcode': order.postcode,
+                    'default_town_or_city': order.town_or_city,
+                    'default_street_address1': order.street_address1,
+                    'default_street_address2': order.street_address2,
+                    'default_county': order.county,
+                }
+                user_profile_form = UserProfileForm(profile_data, instance=profile)
+                if user_profile_form.is_valid():
+                    user_profile_form.save()
+
+        messages.success(request, f'Order successfully processed! Your order number is {order_number}. A confirmation email will be sent to {order.email}.')
+
+        if 'cart' in request.session:
+            del request.session['cart']
+
         template = 'checkout/checkout_success.html'
         context = {
             'order': order,
         }
+
         return render(request, template, context)
-    except Order.DoesNotExist:
-        messages.error(request, f'Order with number {order_number} does not exist.')
-        return redirect(reverse('shop'))
